Add data file to describe find-support form
The aim here is to create a simple YAML representation of a form, that
is absent of any content (Content is abstracted to locales using the
same keys).

This file is used here to:
- Describe the order of questions (order of YAML == order of questions)
- Describe the type of question (eg radio, will be used in view layer to
  select the correct component)
- Describe the validation requirements (can the user skip the question /
  how are answers requried, logic will vary depending on question type)
- describe the availble options

We can also extend this to:
- provide dependent question criteira. Eg by concatinating a previous
  question with a unique option we can provide logic to skip/show a
question based on a previous answer eg: "are_you_a_key_worker_not_sure"
- we could also use this to outline more complex validation logic, as
  data. Required could default to basic checks as set out in the
govuk_publishing components. More complex validation could be an array
which describes ordered validation checks we will need to call, eg:
[required, is_valid_NHS_number]